requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
SQLAlchemy==2.0.43
pymysql==1.0.3
requests==2.32.4
python-dotenv==1.1.1
pillow==11.3.0
